+++ src/main/java/kitchenpos/external/infra/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.external.domain;
+package kitchenpos.external.infra;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
