+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -10,7 +10,6 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
-import kitchenpos.menu_group.domain.MenuGroup;
 
 import java.math.BigDecimal;
 import java.util.List;
