+++ src/main/java/kitchenpos/menu/infra/JpaMenuGroupRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu_group.infra;
+package kitchenpos.menu.infra;
 
-import kitchenpos.menu_group.domain.MenuGroup;
-import kitchenpos.menu_group.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
