+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -12,7 +12,6 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
-import kitchenpos.order_table.domain.OrderTable;
 
 import java.time.LocalDateTime;
 import java.util.List;
