+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -3,7 +3,7 @@ package kitchenpos.product.application;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.external.domain.PurgomalumClient;
+import kitchenpos.external.infra.PurgomalumClient;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
