+++ src/test/java/kitchenpos/external/infra/FakeKitchenridersClient.java
@@ -1,7 +1,5 @@
 package kitchenpos.external.infra;
 
-import kitchenpos.external.domain.KitchenridersClient;
-
 import java.math.BigDecimal;
 import java.util.UUID;
