+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -5,12 +5,12 @@ import kitchenpos.menu.infra.InMemoryMenuRepository;
 import kitchenpos.product.infra.InMemoryProductRepository;
 import kitchenpos.external.infra.FakePurgomalumClient;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu_group.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.external.domain.PurgomalumClient;
+import kitchenpos.external.infra.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
