+++ src/test/java/kitchenpos/menu_group/infra/InMemoryMenuGroupRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu_group.infra;
 
-import kitchenpos.menu_group.domain.MenuGroup;
-import kitchenpos.menu_group.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
 
 import java.util.ArrayList;
 import java.util.HashMap;
