+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -6,8 +6,8 @@ import kitchenpos.order_table.infra.InMemoryOrderTableRepository;
 import kitchenpos.external.infra.FakeKitchenridersClient;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.*;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
