+++ src/test/java/kitchenpos/order/infra/InMemoryOrderRepository.java
@@ -3,7 +3,7 @@ package kitchenpos.order.infra;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order_table.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 import java.util.ArrayList;
 import java.util.HashMap;
