+++ src/test/java/kitchenpos/order_table/application/OrderTableServiceTest.java
@@ -1,11 +1,12 @@
 package kitchenpos.order_table.application;
 
+import kitchenpos.order.application.OrderTableService;
 import kitchenpos.order.infra.InMemoryOrderRepository;
 import kitchenpos.order_table.infra.InMemoryOrderTableRepository;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
