+++ src/test/java/kitchenpos/order_table/infra/InMemoryOrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.order_table.infra;
 
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 
 import java.util.ArrayList;
 import java.util.HashMap;
